+++ src/lib/index.ts
@@ -1,20 +1,21 @@
 export * from './apollo'
 export * from './assert'
 export * from './assetPath'
-export * from './rgba'
 export * from './componentLevelHelpers'
 export * from './cookies'
-export * from './envHelpers'
-export * from './magic'
-export * from './session'
 export * from './CourierNotification'
+export * from './envHelpers'
+export * from './Intercom'
+export * from './magic'
+export * from './payments'
+export * from './rgba'
+export * from './session'
+export * from './useFormLocalStorage'
+export * from './useLoggedInUser'
 export * from './usePreventTabClose'
 export * from './useReadContract'
 export * from './utilsCrypto'
 export * from './utilsServer'
-export * from './payments'
-export * from './useLoggedInUser'
-export * from './Intercom'
 
 export * from './auth'
 /* TODO: There is an error where, if AuthForm is exported before some of these other functions
